refactor: Numbers 변환 기능 제거
